+++ src/RcppExports.cpp
@@ -326,6 +326,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// renumber_tips_to
+SEXP renumber_tips_to(Rcpp::List trees, const Rcpp::CharacterVector target);
+RcppExport SEXP _TreeTools_renumber_tips_to(SEXP treesSEXP, SEXP targetSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::List >::type trees(treesSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::CharacterVector >::type target(targetSEXP);
+    rcpp_result_gen = Rcpp::wrap(renumber_tips_to(trees, target));
+    return rcpp_result_gen;
+END_RCPP
+}
 // cpp_edge_to_splits
 Rcpp::RawMatrix cpp_edge_to_splits(const Rcpp::IntegerMatrix& edge, const Rcpp::IntegerVector& order, const Rcpp::IntegerVector& nTip);
 RcppExport SEXP _TreeTools_cpp_edge_to_splits(SEXP edgeSEXP, SEXP orderSEXP, SEXP nTipSEXP) {
@@ -562,6 +574,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_TreeTools_node_depth_unrooted", (DL_FUNC) &_TreeTools_node_depth_unrooted, 4},
     {"_TreeTools_path_lengths", (DL_FUNC) &_TreeTools_path_lengths, 3},
     {"_TreeTools_renumber_tips_batch", (DL_FUNC) &_TreeTools_renumber_tips_batch, 4},
+    {"_TreeTools_renumber_tips_to", (DL_FUNC) &_TreeTools_renumber_tips_to, 2},
     {"_TreeTools_cpp_edge_to_splits", (DL_FUNC) &_TreeTools_cpp_edge_to_splits, 3},
     {"_TreeTools_duplicated_splits", (DL_FUNC) &_TreeTools_duplicated_splits, 2},
     {"_TreeTools_mask_splits", (DL_FUNC) &_TreeTools_mask_splits, 1},